EEPROM Programmer
Using the Teensy 3.1 to program 28C16 and 28C64 EEPROMS.  Currently the
schematics are for a 28C64 (VCC, A12 and WE are different).  I built a
new breadboard for the 28C16.  The control and address lines on the
Teensy are the same but the wiring will need to be modified for the chip
you are using.  I only plan on using a breadboard...